+++ apps/hono-example/CHANGELOG.md
@@ -1,5 +1,11 @@
 # hono-example
 
+## 0.0.24
+
+### Patch Changes
+
+- @better-i18n/server@0.4.9
+
 ## 0.0.23
 
 ### Patch Changes