+++ tests/expo-native/CHANGELOG.md
@@ -1,5 +1,11 @@
 # expo-native
 
+## 1.0.35
+
+### Patch Changes
+
+- @better-i18n/expo@0.9.2
+
 ## 1.0.34
 
 ### Patch Changes